client/requirements.txt
analytics-python==1.2.9
appnope==0.1.0
arrow==0.12.1
astroid==1.6.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
binaryornot==0.4.4
certifi==2018.4.16
configparser==3.5.0
decorator==4.3.0
defusedxml==0.5.0
docutils==0.14
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
google==2.0.1
googleapis-common-protos==1.5.3
grpc-django==0.3.dev0
grpc-pytools==0.1.2
grpcio==1.12.1
grpcio-tools==1.12.1
idna==2.7
MarkupSafe==1.0
marshmallow==2.15.3
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.6.0
Pygments==2.2.0
pykerberos==1.2.1
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.11
requests==2.19.1
scandir==1.7
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snapshottest==0.5.1
suit==2.0.2
Unidecode==1.0.22
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
whichcraft==0.4.1
wrapt==1.10.11
